tiktok-api/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
Markdown==3.4.3
numpy==1.25.0
opencv-python==4.8.0.74
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
PyJWT==2.7.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
random-user-agent==1.0.1
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
undetected-chromedriver==3.5.0
urllib3==2.0.3
webdriver-manager==3.8.6
websockets==11.0.3
wsproto==1.2.0
